feature/BB-149: 안전결제 방식 적용하여 꿀 사용 로직 개선

+++ payment-service/src/main/java/com/lgcns/bebee/payment/domain/entity/HoneyHistory.java
@@ -24,7 +24,7 @@ public class HoneyHistory extends BaseTimeEntity {
     private Long targetMemberId;
 
     @Column(nullable = false)
-    private Integer amount;
+    private Long amount; // 원 딘위
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
@@ -33,7 +33,7 @@ public class HoneyHistory extends BaseTimeEntity {
     public static HoneyHistory create(
             HoneyWallet honeyWallet,
             Long targetMemberId,
-            Integer amount,
+            Long amount,
             HoneyHistoryType type
     ) {
         HoneyHistory honeyHistory = new HoneyHistory();
@@ -44,4 +44,6 @@ public class HoneyHistory extends BaseTimeEntity {
 
         return honeyHistory;
     }
+
+
 }
